requirements.txt
praw
python-dotenv
pytest
openai